# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple REST API for managing spy cats and missions",
        "title": "Spy Cats API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/missions/{mid}/targets/{tid}/note": {
            "put": {
                "description": "Update the notes of a target",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Update target notes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mission ID",
                        "name": "mid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target notes",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/missions/{mission_id}/targets/{target_id}/complete": {
            "post": {
                "description": "Mark a specific target as completed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Mark target completed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mission ID",
                        "name": "mission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target ID",
                        "name": "target_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}